# gateway-service/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.BalanceResponse:
    properties:
      balance:
        type: number
      user_id:
        type: string
    type: object
  dto.BalanceResponseApi:
    properties:
      data:
        $ref: '#/definitions/dto.BalanceResponse'
      message:
        example: Get data success
        type: string
      status_code:
        example: 200
        type: integer
    required:
    - message
    - status_code
    type: object
...
  dto.BookOrderItem:
    properties:
      book_id:
        type: string
      quantity:
        type: integer
    type: object
...
  dto.CreateTransactionRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.BookOrderItem'
        type: array
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Pesan kesalahan yang deskriptif
        type: string
      message:
        example: Internal Server Error
        type: string
      status_code:
        example: 401
        type: integer
    required:
    - message
    - status_code
    type: object
...
  dto.TopUpRequest:
    properties:
      amount:
        type: number
      method:
        type: string
    required:
    - amount
    - method
    type: object
  dto.TopUpResponse:
    properties:
      amount:
        example: 50000
        type: number
      created_at:
        example: "2025-07-25T15:00:00Z"
        type: string
      method:
        example: credit_card
        type: string
      status:
        example: success
        type: string
      transaction_id:
        example: trx123
        type: string
      user_id:
        example: user123
        type: string
    type: object
  dto.TopUpResponseDto:
    properties:
      data:
        $ref: '#/definitions/dto.TopUpResponse'
      message:
        example: Create data success
        type: string
      status_code:
        example: 201
        type: integer
    required:
    - message
    - status_code
    type: object
  dto.TransactionDetailResponse:
    properties:
      book_id:
        type: string
      price_per_unit:
        type: number
      quantity:
        type: integer
    type: object
  dto.TransactionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TransactionResponse'
        type: array
      message:
        example: Create data success
        type: string
      status_code:
        example: 200
        type: integer
    required:
    - message
    - status_code
    type: object
  dto.TransactionResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/dto.TransactionDetailResponse'
        type: array
      status:
        type: string
      total_amount:
        type: number
      transaction_date:
        type: string
      transaction_id:
        type: string
      user_id:
        type: string
    type: object
  dto.TransactionResponseApi:
    properties:
      data:
        $ref: '#/definitions/dto.TransactionResponse'
      message:
        example: Create data success
        type: string
      status_code:
        example: 201
        type: integer
    required:
    - message
    - status_code
    type: object
...
host: 34.101.226.106:8000
info:
  contact:
    email: dev@booktopia.local
    name: Booktopia Dev Team
  description: Gateway service yang mengelola dan meneruskan permintaan ke berbagai
    microservice di ekosistem Booktopia, termasuk auth-service, book-service, gifting-service,
    transaction-service, dan wallet-service.
  termsOfService: http://swagger.io/terms/
  title: Booktopia Gateway API
  version: "1.0"
paths:
...
  /transactions:
    get:
      description: Mengambil daftar semua transaksi berdasarkan user_id dari token
        JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil semua transaksi milik user
      tags:
      - Gateway - Transaction
    post:
      consumes:
      - application/json
      description: Menerima list item buku dari user dan membuat transaksi baru (via
        gRPC ke transaction-service).
      parameters:
      - description: Data transaksi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TransactionResponseApi'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat transaksi pembelian buku
      tags:
      - Gateway - Transaction
  /wallet/balance:
    get:
      description: Mengambil saldo wallet berdasarkan user_id dari token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BalanceResponseApi'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil saldo wallet user
      tags:
      - Gateway - Wallet
  /wallet/topup:
    post:
      consumes:
      - application/json
      description: Melakukan top up saldo wallet berdasarkan request user
      parameters:
      - description: Detail top up saldo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TopUpResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top up saldo wallet
      tags:
      - Gateway - Wallet
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"